feat: HookableResource - endpoints as hooks
BREAKING CHANGE: Removed Resource.useFetchInit(). Use HookableResource
instead
